+++ wasm/example/browser/src/index.js
@@ -2,11 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import CelParserComponent from "./CELParser";
+import SuperscriptParserComponent from "./SuperscriptParser";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <CelParserComponent />
+    <SuperscriptParserComponent />
   </React.StrictMode>
 );